Test abstract filter class
