+++ src/main/java/seedu/address/logic/commands/Filter.java
@@ -1,0 +1,44 @@
+package seedu.address.logic.commands;
+
+import seedu.address.commons.util.ToStringBuilder;
+import seedu.address.logic.Messages;
+import seedu.address.model.Model;
+import seedu.address.model.person.Person;
+
+import java.util.function.Predicate;
+
+import static java.util.Objects.requireNonNull;
+
+public abstract class Filter extends Command {
+    Predicate<Person> predicate;
+
+    @Override
+    public CommandResult execute(Model model) {
+        requireNonNull(model);
+        model.updateFilteredPersonList(predicate);
+        return new CommandResult(
+                String.format(Messages.MESSAGE_PERSONS_LISTED_OVERVIEW, model.getFilteredPersonList().size()));
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        if (other == this) {
+            return true;
+        }
+
+        // instanceof handles nulls
+        if (!(other instanceof Filter)) {
+            return false;
+        }
+
+        Filter otherFilter = (Filter) other;
+        return predicate.equals(otherFilter.predicate);
+    }
+
+    @Override
+    public String toString() {
+        return new ToStringBuilder(this)
+                .add("predicate", predicate)
+                .toString();
+    }
+}
