+++ src/main/java/seedu/address/logic/commands/FilterCommand.java
@@ -5,21 +5,19 @@ import static java.util.Objects.requireNonNull;
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.Messages;
 import seedu.address.model.Model;
-import seedu.address.model.person.PersonHasTagPredicate;
+import seedu.address.model.person.predicates.PersonHasTagPredicate;
 
 /**
  * Sorts and lists all persons in address book who are tagged by any of the argument keywords.
  * Keyword matching is case-sensitive.
  */
-public class FilterCommand extends Command {
+public class FilterCommand extends Filter {
     public static final String COMMAND_WORD = "filter";
 
     public static final String MESSAGE_USAGE = COMMAND_WORD + ": Filters for all persons whose tags include any of "
             + "the specified keywords and displays them as a list with index numbers.\n"
             + "Parameters: KEYWORD [MORE_KEYWORDS]...\n"
             + "Example: " + COMMAND_WORD + " friend teacher student";
-
-    private final PersonHasTagPredicate predicate;
 
     /**
      * Returns a new FilterCommand object that takes in a PersonHasTagPredicate
@@ -28,35 +26,5 @@ public class FilterCommand extends Command {
      */
     public FilterCommand(PersonHasTagPredicate predicate) {
         this.predicate = predicate;
-    }
-
-    @Override
-    public CommandResult execute(Model model) {
-        requireNonNull(model);
-        model.updateFilteredPersonList(predicate);
-        return new CommandResult(
-                String.format(Messages.MESSAGE_PERSONS_LISTED_OVERVIEW, model.getFilteredPersonList().size()));
-    }
-
-    @Override
-    public boolean equals(Object other) {
-        if (other == this) {
-            return true;
-        }
-
-        // instanceof handles nulls
-        if (!(other instanceof FilterCommand)) {
-            return false;
-        }
-
-        FilterCommand otherFilterCommand = (FilterCommand) other;
-        return predicate.equals(otherFilterCommand.predicate);
-    }
-
-    @Override
-    public String toString() {
-        return new ToStringBuilder(this)
-                .add("predicate", predicate)
-                .toString();
     }
 }
