+++ src/main/java/seedu/address/model/person/predicates/PersonAvailableOnDayPredicate.java
@@ -1,0 +1,4 @@
+package seedu.address.model.person.predicates;
+
+public class PersonAvailableOnDayPredicate {
+}
